backend/requirements.txt
Django==2.0.7
djangorestframework==3.8.2
markdown==2.6.11
django-filter==2.0.0
django-cors-headers==2.4.0
djongo==1.2.29
djangorestframework_jwt==1.11.0
pyjwt==1.6.4
python-jose==3.0.0
pymysql==0.9.2
jenkinsapi==0.3.6
gunicorn==19.9.0
uwsgi==2.0.17.1
whitenoise==3.3.1